+++ src/data/topol/graph/csr/impls/mod.rs
@@ -1,0 +1,4 @@
+// devela/src/data/topol/graph/csr/impls/mod.rs
+
+mod array; // __graph_csr_impl_array!
+mod vec; // __graph_csr_impl_vec!